[feat] board, comment request 전부 token 으로 받아서 user 구분

+++ server/src/main/java/com/Petching/petching/board/controller/BoardController.java
@@ -5,7 +5,9 @@ import com.Petching.petching.board.entity.Board;
 import com.Petching.petching.board.mapper.BoardMapper;
 import com.Petching.petching.board.service.BoardService;
 import com.Petching.petching.comment.mapper.CommentMapper;
-import com.Petching.petching.login.jwt.service.JwtService;
+import com.Petching.petching.global.exception.BusinessLogicException;
+import com.Petching.petching.global.exception.ExceptionCode;
+import com.Petching.petching.login.oauth.userInfo.JwtToken;
 import com.Petching.petching.login.service.UserLoginService;
 import com.Petching.petching.response.MultiResponse;
 import com.Petching.petching.response.PageInfo;
@@ -13,6 +15,7 @@ import com.Petching.petching.response.SingleResponse;
 import com.Petching.petching.user.entity.User;
 import com.Petching.petching.user.service.UserService;
 import com.google.gson.Gson;
+import com.nimbusds.jose.util.IntegerUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -20,15 +23,17 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.Nullable;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
+import org.thymeleaf.util.StringUtils;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.net.URI;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Collectors;
 
 
 @RestController
@@ -40,16 +45,16 @@ public class BoardController {
     private final BoardMapper mapper;
     private final UserService userService;
 
-    private final JwtService jwtService;
-
     private final CommentMapper commentMapper;
 
-    public BoardController(BoardService boardService, BoardMapper mapper, UserService userService, JwtService jwtService, CommentMapper commentMapper) {
+    private final JwtToken jwtToken;
+
+    public BoardController(BoardService boardService, BoardMapper mapper, UserService userService, CommentMapper commentMapper, JwtToken jwtToken) {
         this.boardService = boardService;
         this.mapper = mapper;
         this.userService = userService;
-        this.jwtService = jwtService;
         this.commentMapper = commentMapper;
+        this.jwtToken = jwtToken;
     }
 /**
  * 아래는 글쓰는 요청 시 post Dto 와 File 을 한 번에 받을 수 있는 메서드
@@ -87,8 +92,8 @@ public class BoardController {
     public ResponseEntity postBoard(@Valid @RequestBody BoardDto.Post requestBody,
                                     @RequestHeader("Authorization") String authorization){
 
-        Long userId = jwtService.getUserIdFromAuthHeader(authorization);
-        User user = userService.findUser(userId);
+        authorization = authorization.replaceAll("Bearer ","");
+        User user = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
 
         Board board = mapper.boardPostDtoToBoard(requestBody);
         board.setUser(user);
@@ -100,6 +105,7 @@ public class BoardController {
                 .build().toUri();
 
         return ResponseEntity.created(uri).build();
+
     }
 
 
@@ -107,15 +113,24 @@ public class BoardController {
      * request header 로 요청받아서 유저를 검증하는 방법으로 리팩토링 필요
      * */
     @PatchMapping("/{boardId}")
-    public ResponseEntity patchBoard(@PathVariable("boardId") long id, @RequestBody BoardDto.Patch requestBody){
+    public ResponseEntity patchBoard(@PathVariable("boardId") long boardId,
+                                     @RequestBody BoardDto.Patch requestBody,
+                                     @RequestHeader("Authorization") String authorization){
 
-        Board board = mapper.boardPatchDtoToBoard(requestBody);
-        board.setBoardId(id);
+        authorization = authorization.replaceAll("Bearer ","");
+        User requestUser = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
 
-        Board response = boardService.updateBoard(board);
+        Board originalBoard = boardService.findBoardByMK(boardId);
+        User writer = originalBoard.getUser();
 
-        return new ResponseEntity<>(new SingleResponse<>(mapper.boardToBoardDetailDto(response)), HttpStatus.OK);
+        if(Objects.equals(writer.getUserId(), requestUser.getUserId())){
+            Board board = mapper.boardPatchDtoToBoard(requestBody);
+            board.setBoardId(boardId);
+            Board response = boardService.updateBoard(board);
+            return new ResponseEntity<>(new SingleResponse<>(mapper.boardToBoardDetailDto(response)), HttpStatus.OK);
+        }
 
+        throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
     }
 
     /* 글을 삭제하는 API
@@ -123,49 +138,45 @@ public class BoardController {
      * */
 
     @DeleteMapping("/{boardId}")
-    public ResponseEntity deleteBoard(@PathVariable("boardId") long boardId){
+    public ResponseEntity deleteBoard(@PathVariable("boardId") long boardId,
+                                      @RequestHeader("Authorization") String authorization){
 
-        boardService.deleteBoard(boardId);
+        authorization = authorization.replaceAll("Bearer ","");
+        User requestUser = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
 
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Board originalBoard = boardService.findBoardByMK(boardId);
+        User writer = originalBoard.getUser();
 
+        if(Objects.equals(writer.getUserId(), requestUser.getUserId())){
+            boardService.deleteBoard(boardId);
+
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
     }
 
     // 전체 게시글 목록 조회
     @GetMapping
     public ResponseEntity getBoards(
             @RequestParam(defaultValue = "1") int page,
-            @RequestParam(defaultValue = "10") int size
+            @RequestParam(defaultValue = "10") int size,
+            @Nullable @RequestHeader("Authorization") String authorization
     ){
 
         Pageable pageable = PageRequest.of(page-1, size);
         Page<Board> boardPage = boardService.findBoards(pageable);
 
-        /** board service 에 넣을 것
-         * User 가 해당 게시물을 좋아하는 지 확인
-         * user 엔티티에서 LikedBoardList 추가 필요
-         * User user = userService.findUser(userId);
-         * List<Long> userLiked = user.getLikedBoardList()
-         * .stream()
-         * .filter( boardId -> boardId == id)
-         * .collect(Collectors.toList());
-         * if(!userLiked.isEmpty())
-         * board.setCheckLike(true);
-         * */
-
-
         List<BoardDto.Response> response = mapper.boardPageToBoardResponseListDto(boardPage);
 
-        // 페이지 정보 헤더에 추가
-        PageInfo pageInfo = new PageInfo(
-                boardPage.getNumber(),
-                boardPage.getSize(),
-                boardPage.getTotalElements(),
-                boardPage.getTotalPages()
-        );
+        if(!StringUtils.isEmpty(authorization)){
+            authorization = authorization.replaceAll("Bearer ","");
+            User requestUser = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
 
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("X-Page-Info", new Gson().toJson(pageInfo));
+            List<Long> userLiked = requestUser.getLikedBoardList();
+
+            boardService.extracted(response, userLiked);
+        }
 
         return new ResponseEntity<>(new MultiResponse<>(response, boardPage), HttpStatus.OK);
 
@@ -173,25 +184,20 @@ public class BoardController {
 
     // 특정 게시글 조회
     @GetMapping("/{boardId}")
-    public ResponseEntity getBoard(@PathVariable("boardId") long id, @RequestParam long userId){
+    public ResponseEntity getBoard(@PathVariable("boardId") long id,
+                                   @Nullable @RequestHeader("Authorization") String authorization){
 
         Board board = boardService.findBoardByMK(id);
-
-        /**
-         * board service 에 넣을 것
-         * User 가 해당 게시물을 좋아하는 지 확인
-         * user 엔티티에서 LikedBoardList 추가 필요
-         * User user = userService.findUser(userId);
-         * List<Long> userLiked = user.getLikedBoardList()
-         * .stream()
-         * .filter( boardId -> boardId == id)
-         * .collect(Collectors.toList());
-         * if(!userLiked.isEmpty())
-         * board.setCheckLike(true);
-         * */
-
         BoardDto.Detail response = mapper.boardToBoardDetailDto(board);
         response.setComments(commentMapper.commentListToCommentResponseListDto(board.getComments()));
+
+        if(!StringUtils.isEmpty(authorization)){
+            authorization = authorization.replaceAll("Bearer ","");
+            User requestUser = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
+
+            if(requestUser.getLikedBoardList().contains(board.getBoardId()))
+                response.setCheckLike(true);
+        }
 
         return new ResponseEntity<>( new SingleResponse<>(response), HttpStatus.OK);
 
@@ -213,20 +219,19 @@ public class BoardController {
     *  userId 는 request header 로 요청받아서 유저를 찾는 방법으로 리팩토링 필요
     * */
     @PostMapping("/{boardId}/like")
-    public ResponseEntity updateLike(@PathVariable("boardId") long boardId, @RequestParam long userId){
+    public ResponseEntity updateLike(@PathVariable("boardId") long boardId,
+                                     @RequestHeader("Authorization") String authorization){
 
-        Board updatedBoard = boardService.updateBoardLike(boardId);
+        authorization = authorization.replaceAll("Bearer ","");
+        User requestUser = userService.findUser(jwtToken.extractUserIdFromToken(authorization));
+        requestUser.addLikedBoard(boardId);
+        userService.updatedByBoardLike(requestUser);
 
+        boardService.updateBoardLike(boardId, requestUser);
 
-        /**
-         *  User 가 좋아하는 게시글 저장
-         *  user service 에서 구현 필요.
-         *  User user = userService.findUser(userId);
-         *  user.addLikedBoard(updatedBoard.getBoardId());
-         *  userService.updateFromBoard(user);
-         *
-         * */
 
         return new ResponseEntity( HttpStatus.OK);
     }
+
+
 }
